+++ ServicioReparaciones-web/src/main/java/com/servicio/reparaciones/web/bean/SalidaBean.java
@@ -7,11 +7,13 @@ package com.servicio.reparaciones.web.bean;
 
 import com.servicio.reparaciones.modelo.nosql.Articulo;
 import com.servicio.reparaciones.modelo.nosql.Bodega;
+import com.servicio.reparaciones.modelo.nosql.Inventario;
 import com.servicio.reparaciones.modelo.nosql.Salida;
 import com.servicio.reparaciones.modelo.nosql.Tecnico;
 import com.servicio.reparaciones.modelo.nosql.Usuario;
 import com.servicio.reparaciones.servicio.ArticuloService;
 import com.servicio.reparaciones.servicio.BodegaService;
+import com.servicio.reparaciones.servicio.InventarioService;
 import com.servicio.reparaciones.servicio.SalidaService;
 import com.servicio.reparaciones.servicio.TecnicoServicio;
 import com.servicio.reparaciones.servicio.UsuarioServicio;
@@ -20,6 +22,8 @@ import com.servicio.reparaciones.web.util.FacesUtil;
 import com.servicio.reparaciones.web.util.SessionUtil;
 import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.Collections;
+import java.util.HashSet;
 import java.util.List;
 import javax.annotation.PostConstruct;
 import javax.faces.event.ActionEvent;
@@ -44,6 +48,7 @@ public class SalidaBean implements ImethodsBean, Serializable {
     private List<Salida> filtered;
     private List<Articulo> articulos;
     private List<Articulo> filteredArticulos;
+    private Double cantidadExistente;
 
     private Usuario usuario;
 
@@ -54,6 +59,8 @@ public class SalidaBean implements ImethodsBean, Serializable {
     @Inject
     private SalidaService salidaService;
     @Inject
+    private InventarioService inventarioService;
+    @Inject
     private TecnicoServicio tecnicoService;
     @Inject
     private UsuarioServicio usarioService;
@@ -62,8 +69,10 @@ public class SalidaBean implements ImethodsBean, Serializable {
     public void init() {
         this.nuevo = new Salida();
         this.selected = new Salida();
+        this.cantidadExistente = 0.0;
         this.filtered = null;
         this.salidas = this.salidaService.ObtenerListaSalidas(1);
+        Collections.reverse(this.salidas);
         this.articulos = new ArrayList<>();
         this.usuario = new Usuario();
         this.usuario.setCodigo(SessionUtil.sessionVarNumeric("codigo"));
@@ -80,13 +89,24 @@ public class SalidaBean implements ImethodsBean, Serializable {
         Bodega bodega = this.bodegaService.findByCodigo(this.nuevo.getBodega());
         this.nuevo.setBodega(bodega);
         this.nuevo.setCode("SROUT" + (this.salidaService.generatedCodigo() - 1));
-        Boolean exito = this.salidaService.insert(this.nuevo);
-        if (exito) {
-            FacesUtil.addMessageInfo("Se ha guardado con exito.");
-            this.init();
+        if (this.nuevo.getBodega().getUsername().equals(this.nuevo.getUsername())) {
+            this.cantidadExistente = this.inventarioService.
+                    CalcularExistenciasBodegaArticulo(
+                            this.nuevo.getBodega(), this.nuevo.getArticulo());
+            if (this.cantidadExistente >= this.nuevo.getCantidad()) {
+                Boolean exito = this.salidaService.insert(this.nuevo);
+                if (exito) {
+                    FacesUtil.addMessageInfo("Se ha guardado con exito.");
+                    this.init();
+                } else {
+                    FacesUtil.addMessageError(null, "No se ha guardado.");
+                    this.init();
+                }
+            } else {
+                FacesUtil.addMessageInfo("No hay unidades disponibles en bodega.");
+            }
         } else {
-            FacesUtil.addMessageError(null, "No se ha guardado.");
-            this.init();
+            FacesUtil.addMessageInfo("Usted no está autorizado");
         }
     }
 
@@ -101,13 +121,24 @@ public class SalidaBean implements ImethodsBean, Serializable {
             this.selected.setArticulo(articulo);
             Bodega bodega = this.bodegaService.findByCodigo(this.selected.getBodega());
             this.selected.setBodega(bodega);
-            Boolean exito = this.salidaService.update(this.selected);
-            if (exito) {
-                FacesUtil.addMessageInfo("Se ha modifcado con exito.");
-                this.init();
+            if (this.nuevo.getBodega().getUsername().equals(this.nuevo.getUsername())) {
+                this.cantidadExistente = this.inventarioService.
+                        CalcularExistenciasBodegaArticulo(
+                                this.nuevo.getBodega(), this.nuevo.getArticulo());
+                if (this.cantidadExistente >= this.nuevo.getCantidad()) {
+                    Boolean exito = this.salidaService.update(this.selected);
+                    if (exito) {
+                        FacesUtil.addMessageInfo("Se ha modifcado con exito.");
+                        this.init();
+                    } else {
+                        FacesUtil.addMessageError(null, "No se ha modifcado con exito..");
+                        this.init();
+                    }
+                } else {
+                    FacesUtil.addMessageInfo("No hay unidades disponibles en bodega.");
+                }
             } else {
-                FacesUtil.addMessageError(null, "No se ha modifcado con exito..");
-                this.init();
+                FacesUtil.addMessageInfo("Usted no está autorizado.");
             }
         } else {
             FacesUtil.addMessageInfo("Seleccione un registro.");
@@ -119,13 +150,17 @@ public class SalidaBean implements ImethodsBean, Serializable {
         if (this.selected != null) {
             this.usuario = this.usarioService.findByCodigo(this.usuario);
             this.selected.setUsername(this.usuario);
-            Boolean exito = this.salidaService.deleteFlag(this.selected);
-            if (exito) {
-                FacesUtil.addMessageInfo("Se ha eliminado con exito.");
-                this.init();
+            if (this.nuevo.getBodega().getUsername().equals(this.nuevo.getUsername())) {
+                Boolean exito = this.salidaService.deleteFlag(this.selected);
+                if (exito) {
+                    FacesUtil.addMessageInfo("Se ha eliminado con exito.");
+                    this.init();
+                } else {
+                    FacesUtil.addMessageError(null, "No se ha eliminado con exito..");
+                    this.init();
+                }
             } else {
-                FacesUtil.addMessageError(null, "No se ha eliminado con exito..");
-                this.init();
+                FacesUtil.addMessageInfo("Usted no está autorizado.");
             }
         } else {
             FacesUtil.addMessageInfo("Seleccione un registro.");
@@ -136,6 +171,9 @@ public class SalidaBean implements ImethodsBean, Serializable {
         Articulo art = (Articulo) event.getObject();
         if (art != null) {
             this.nuevo.setArticulo(art);
+            this.cantidadExistente = this.inventarioService.
+                    CalcularExistenciasBodegaArticulo(
+                            this.nuevo.getBodega(), this.nuevo.getArticulo());
         }
     }
 
@@ -143,17 +181,62 @@ public class SalidaBean implements ImethodsBean, Serializable {
         Articulo art = (Articulo) event.getObject();
         if (art != null && this.selected != null) {
             this.selected.setArticulo(art);
+            this.cantidadExistente = this.inventarioService.
+                    CalcularExistenciasBodegaArticulo(
+                            this.selected.getBodega(), this.selected.getArticulo());
         }
     }
 
-    public void loadBodega() {
+    public void loadArticulos() {
         Bodega bodega = this.bodegaService.findByCodigo(this.nuevo.getBodega());
-        
-        
+        this.nuevo.setBodega(bodega);
+        List<Inventario> narticulos = this.inventarioService.findByBodega(bodega);
+        this.articulos = new ArrayList<>();
+        if (narticulos != null && !narticulos.isEmpty()) {
+            for (Inventario in : narticulos) {
+                this.articulos.add(in.getArticulo());
+            }
+            HashSet<Articulo> hashSet = new HashSet<Articulo>(this.articulos);
+            this.articulos.clear();
+            this.articulos.addAll(hashSet);
+        }
     }
 
-    public void loadSelectBodega() {
+    public void loadSelectArticulos() {
         Bodega bodega = this.bodegaService.findByCodigo(this.selected.getBodega());
+        this.selected.setBodega(bodega);
+        List<Inventario> narticulos = this.inventarioService.findByBodega(bodega);
+        if (narticulos != null && !narticulos.isEmpty()) {
+            this.articulos = new ArrayList<>();
+            for (Inventario in : narticulos) {
+                this.articulos.add(in.getArticulo());
+            }
+            HashSet<Articulo> hashSet = new HashSet<Articulo>(this.articulos);
+            this.articulos.clear();
+            this.articulos.addAll(hashSet);
+            this.cantidadExistente = this.inventarioService.
+                    CalcularExistenciasBodegaArticulo(
+                            this.selected.getBodega(), this.selected.getArticulo());
+        }
+    }
+
+    public void loadSelectArticulos(Salida selected) {
+        this.selected = selected;
+        Bodega bodega = this.bodegaService.findByCodigo(this.selected.getBodega());
+        this.selected.setBodega(bodega);
+        List<Inventario> narticulos = this.inventarioService.findByBodega(bodega);
+        if (narticulos != null && !narticulos.isEmpty()) {
+            this.articulos = new ArrayList<>();
+            for (Inventario in : narticulos) {
+                this.articulos.add(in.getArticulo());
+            }
+            HashSet<Articulo> hashSet = new HashSet<Articulo>(this.articulos);
+            this.articulos.clear();
+            this.articulos.addAll(hashSet);
+            this.cantidadExistente = this.inventarioService.
+                    CalcularExistenciasBodegaArticulo(
+                            this.selected.getBodega(), this.selected.getArticulo());
+        }
     }
 
     public Salida getNuevo() {
@@ -202,6 +285,14 @@ public class SalidaBean implements ImethodsBean, Serializable {
 
     public void setFilteredArticulos(List<Articulo> filteredArticulos) {
         this.filteredArticulos = filteredArticulos;
+    }
+
+    public Double getCantidadExistente() {
+        return cantidadExistente;
+    }
+
+    public void setCantidadExistente(Double cantidadExistente) {
+        this.cantidadExistente = cantidadExistente;
     }
 
 }
